Renaming plugins to "team" - (Task #195)
---
Task #195: Git support for VirSat projects
