+++ de.dlr.sc.virsat.svn.test/build.properties
@@ -2,5 +2,4 @@ source.. = src/
 output.. = target/classes/
 bin.includes = META-INF/,\
                .,\
-               resources/,\
                about.html
