+++ de.dlr.sc.virsat.svn.ui/META-INF/MANIFEST.MF
@@ -1,9 +1,9 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: VirSat Core Simple SVN Connector UI
+Bundle-Name: VirSat Core Simple Team Connector UI
 Bundle-SymbolicName: de.dlr.sc.virsat.team.ui;singleton:=true
 Bundle-Version: 4.9.1.qualifier
-Bundle-Activator: de.dlr.sc.virsat.svn.ui.Activator
+Bundle-Activator: de.dlr.sc.virsat.team.ui.Activator
 Bundle-Vendor: DLR (German Aerospace Center)
 Require-Bundle: org.eclipse.ui,
  de.dlr.sc.virsat.team,
