+++ de.dlr.sc.virsat.svn.ui/plugin.xml
@@ -15,7 +15,7 @@
    <extension
          point="org.eclipse.ui.handlers">
       <handler
-            class="de.dlr.sc.virsat.svn.ui.action.SVNCommitAction"
+            class="de.dlr.sc.virsat.team.ui.action.svn.SVNCommitAction"
             commandId="de.dlr.sc.virsat.team.ui.command.commit">
          <enabledWhen>
             <with
@@ -32,7 +32,7 @@
          </enabledWhen>
       </handler>
       <handler
-            class="de.dlr.sc.virsat.svn.ui.action.SVNUpdateAction"
+            class="de.dlr.sc.virsat.team.ui.action.svn.SVNUpdateAction"
             commandId="de.dlr.sc.virsat.team.ui.command.update">
          <activeWhen>
             <with
@@ -49,7 +49,7 @@
          </activeWhen>
       </handler>
  	  <handler
-            class="de.dlr.sc.virsat.team.ui.git.action.GitCommitAction"
+            class="de.dlr.sc.virsat.team.ui.action.git.GitCommitAction"
             commandId="de.dlr.sc.virsat.team.ui.command.commit">
          <activeWhen>
             <with
@@ -65,7 +65,7 @@
          </activeWhen>
       </handler>
       <handler
-            class="de.dlr.sc.virsat.team.ui.git.action.GitUpdateAction"
+            class="de.dlr.sc.virsat.team.ui.action.git.GitUpdateAction"
             commandId="de.dlr.sc.virsat.team.ui.command.update">
          <activeWhen>
             <with
@@ -123,7 +123,7 @@
    <extension
          point="org.eclipse.core.expressions.propertyTesters">
       <propertyTester
-            class="de.dlr.sc.virsat.svn.ui.propertyTester.VirSatPropertyTesterSharedProject"
+            class="de.dlr.sc.virsat.team.ui.propertyTester.VirSatPropertyTesterSharedProject"
             id="de.dlr.sc.virsat.svn.ui.ProjectShared"
             namespace="de.dlr.sc.virsat.svn.ui"
             properties="svnEnabled, gitEnabled"
