+++ de.dlr.sc.virsat.svn.ui/src/de/dlr/sc/virsat/team/ui/Activator.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.svn.ui;
+package de.dlr.sc.virsat.team.ui;
 
 import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.osgi.framework.BundleContext;
