+++ de.dlr.sc.virsat.svn.ui/src/de/dlr/sc/virsat/team/ui/action/svn/SVNCommitAction.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.svn.ui.action;
+package de.dlr.sc.virsat.team.ui.action.svn;
 
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
@@ -39,7 +39,7 @@ import org.eclipse.ui.handlers.HandlerUtil;
 import de.dlr.sc.virsat.project.editingDomain.VirSatTransactionalEditingDomain;
 import de.dlr.sc.virsat.project.ui.Activator;
 import de.dlr.sc.virsat.project.ui.navigator.util.VirSatSelectionHelper;
-import de.dlr.sc.virsat.svn.ui.dialog.CommitMessageDialog;
+import de.dlr.sc.virsat.team.ui.dialog.CommitMessageDialog;
 
 /**
  * a class to define the commit action
