+++ de.dlr.sc.virsat.svn.ui/src/de/dlr/sc/virsat/team/ui/action/svn/SVNUpdateAction.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.svn.ui.action;
+package de.dlr.sc.virsat.team.ui.action.svn;
 
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
