+++ de.dlr.sc.virsat.svn.ui/src/de/dlr/sc/virsat/team/ui/propertyTester/VirSatPropertyTesterSharedProject.java
@@ -8,7 +8,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package de.dlr.sc.virsat.svn.ui.propertyTester;
+package de.dlr.sc.virsat.team.ui.propertyTester;
 
 import org.eclipse.core.internal.propertytester.ResourcePropertyTester;
 import org.eclipse.core.resources.IProject;
