+++ de.dlr.sc.virsat.svn/META-INF/MANIFEST.MF
@@ -1,13 +1,13 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: VirSat Core Simple SVN Connector
+Bundle-Name: VirSat Core Simple Team Connector
 Bundle-SymbolicName: de.dlr.sc.virsat.team
 Bundle-Version: 4.9.1.qualifier
-Bundle-Activator: de.dlr.sc.virsat.svn.Activator
+Bundle-Activator: de.dlr.sc.virsat.team.Activator
 Bundle-Vendor: DLR (German Aerospace Center)
 Require-Bundle: org.eclipse.core.runtime
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Eclipse-ExtensibleAPI: true
-Export-Package: de.dlr.sc.virsat.svn
+Export-Package: de.dlr.sc.virsat.team
 Automatic-Module-Name: de.dlr.sc.virsat.svn
