+++ de.dlr.sc.virsat.svn/pom.xml
@@ -17,7 +17,7 @@ SPDX-License-Identifier: EPL-2.0
 		<relativePath>../maven/pom.xml</relativePath>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
-	<artifactId>de.dlr.sc.virsat.svn</artifactId>
-	<name>de.dlr.sc.virsat.svn</name>
+	<artifactId>de.dlr.sc.virsat.team</artifactId>
+	<name>de.dlr.sc.virsat.team</name>
 	<packaging>eclipse-plugin</packaging>
 </project>
